refactor(ai): add AI nav section w/experiments
